requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
numpy==1.26.4
pillow==10.2.0
requests==2.31.0
selenium==4.18.1
tensorflow==2.16.1
Werkzeug==3.0.1